Fixed overwriting clipboard on Label's MouseDoubleClick (Title bar of root folder)
Use workaround class for "Clipboard" issue on .Net Windows Forms Label:
On Label MouseDoubleClick the framework will copy the title text into the clipboard.
We avoid this by overriding the Text atrribute and use our own copy.
(https://stackoverflow.com/questions/2519587/is-there-any-way-to-disable-the-double-click-to-copy-functionality-of-a-net-l)

Note: When you have trouble with the Visual Studio Designer not showing the GUI properly, simply build once and reopen the Designer.
This will place the required files into the Designer's cache and becomes able to show the GUI as usual.

Fixes #5 (https://github.com/Hofknecht/SystemTrayMenu/issues/5)