Add protection against DisconnectedItem as sender object in mouse event handlers
Workaround was added in early development stage already:
1f985c53aac8076e807df67778f435201c2bb115
However, it was meant to be no longer occurring, so detection was removed here:
6ebfed5f78efa23fd0bd2c5dbad5221fa723133b
3f09f4018830fc439e32dfc801b021ad521eabe6
3153032110c928a8fddcb4010e5a52b5048b1003

Since it occurred again, we reintroduce this fix at least on all mouse events.
This time, we check for the expected object type rather looking up the item in the list.